Added naive softmax and attention kernels.

+++ transpose/transpose_naive.cu
@@ -1,7 +1,7 @@
 #include <cuda_runtime.h>
 #include <stdio.h>
 
-__global__ void transpose_naive_kernel(long *input, long *output, int row, int col) {
+__global__ void transpose_naive_kernel(float *input, float *output, int row, int col) {
     int i = blockIdx.y * blockDim.y + threadIdx.y;
     int j = blockIdx.x * blockDim.x + threadIdx.x;
     
@@ -11,7 +11,7 @@ __global__ void transpose_naive_kernel(long *input, long *output, int row, int c
 }
 
 // Host function to launch kernel
-extern "C" void transpose_naive_cuda(long *input, long *output, int row, int col) {
+extern "C" void transpose_naive_cuda(float *input, float *output, int row, int col) {
     dim3 block(16, 16);
     dim3 grid((col + block.x - 1) / block.x, (row + block.y - 1) / block.y);
     transpose_naive_kernel<<<grid, block>>>(input, output, row, col);
